SkyBox Update

+++ Code/Engine/Graphics/Material.h
@@ -28,7 +28,7 @@ namespace Engine
 			size_t materialUniformCount;
 			char** materialUniformNames;
 			MaterialUniform * materialUniforms;
-			size_t mapCount;
+			int mapCount;
 			Map* mTextureMaps;
 			Material();
 			Material(char*);
